Remove Install/Unlink PartnerContactPerson use cases and related tests
