--- src/ContactPersons/UseCase/InstallPartnerContactPerson/Handler.php
@@ -1,65 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Bitrix24\Lib\ContactPersons\UseCase\InstallPartnerContactPerson;
-
-use Bitrix24\Lib\ContactPersons\Entity\ContactPerson;
-use Bitrix24\Lib\Services\Flusher;
-use Bitrix24\SDK\Application\Contracts\ApplicationInstallations\Entity\ApplicationInstallationInterface;
-use Bitrix24\SDK\Application\Contracts\ApplicationInstallations\Repository\ApplicationInstallationRepositoryInterface;
-use Bitrix24\SDK\Application\Contracts\ContactPersons\Entity\ContactPersonStatus;
-use Bitrix24\SDK\Application\Contracts\ContactPersons\Repository\ContactPersonRepositoryInterface;
-use Bitrix24\SDK\Application\Contracts\Events\AggregateRootEventsEmitterInterface;
-use Psr\Log\LoggerInterface;
-use Symfony\Component\Uid\Uuid;
-
-readonly class Handler
-{
-    public function __construct(
-        private ApplicationInstallationRepositoryInterface $applicationInstallationRepository,
-        private ContactPersonRepositoryInterface $contactPersonRepository,
-        private Flusher $flusher,
-        private LoggerInterface $logger
-    ) {}
-
-    public function handle(Command $command): void
-    {
-        $this->logger->info('ContactPerson.InstallPartnerContactPerson.start', [
-            'applicationInstallationId' => $command->applicationInstallationId,
-            'bitrix24UserId' => $command->bitrix24UserId,
-        ]);
-
-        /** @var null|AggregateRootEventsEmitterInterface|ApplicationInstallationInterface $applicationInstallation */
-        $applicationInstallation = $this->applicationInstallationRepository->getById($command->applicationInstallationId);
-
-        $uuidV7 = Uuid::v7();
-
-        $contactPerson = new ContactPerson(
-            $uuidV7,
-            ContactPersonStatus::active,
-            $command->fullName,
-            $command->email,
-            null,
-            $command->mobilePhoneNumber,
-            null,
-            $command->comment,
-            $command->externalId,
-            $command->bitrix24UserId,
-            $command->bitrix24PartnerId,
-            $command->userAgentInfo,
-            true
-        );
-
-        $this->contactPersonRepository->save($contactPerson);
-
-        $applicationInstallation->linkBitrix24PartnerContactPerson($uuidV7);
-        $this->applicationInstallationRepository->save($applicationInstallation);
-
-        $this->flusher->flush($contactPerson, $applicationInstallation);
-
-        $this->logger->info('ContactPerson.InstallPartnerContactPerson.finish', [
-            'contact_person_id' => $uuidV7->toRfc4122(),
-        ]);
-    }
-}
