--- tests/Functional/ContactPersons/UseCase/InstallPartnerContactPerson/HandlerTest.php
@@ -1,201 +1,0 @@
-<?php
-
-/**
- * This file is part of the bitrix24-php-lib package.
- *
- * © Maksim Mesilov <mesilov.maxim@gmail.com>
- *
- * For the full copyright and license information, please view the MIT-LICENSE.txt
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Bitrix24\Lib\Tests\Functional\ContactPersons\UseCase\InstallPartnerContactPerson;
-
-use Bitrix24\Lib\ApplicationInstallations\Infrastructure\Doctrine\ApplicationInstallationRepository;
-use Bitrix24\Lib\Bitrix24Accounts\Infrastructure\Doctrine\Bitrix24AccountRepository;
-use Bitrix24\Lib\ContactPersons\UseCase\InstallPartnerContactPerson\Handler;
-use Bitrix24\Lib\ContactPersons\UseCase\InstallPartnerContactPerson\Command;
-use Bitrix24\Lib\ContactPersons\Infrastructure\Doctrine\ContactPersonRepository;
-use Bitrix24\Lib\Services\Flusher;
-use Bitrix24\Lib\Tests\Functional\ApplicationInstallations\Builders\ApplicationInstallationBuilder;
-use Bitrix24\Lib\Tests\Functional\Bitrix24Accounts\Builders\Bitrix24AccountBuilder;
-use Bitrix24\SDK\Application\ApplicationStatus;
-use Bitrix24\SDK\Application\Contracts\ApplicationInstallations\Entity\ApplicationInstallationStatus;
-use Bitrix24\SDK\Application\Contracts\ApplicationInstallations\Events\ApplicationInstallationBitrix24PartnerContactPersonLinkedEvent;
-use Bitrix24\SDK\Application\Contracts\ApplicationInstallations\Events\ApplicationInstallationBitrix24PartnerLinkedEvent;
-use Bitrix24\SDK\Application\Contracts\ApplicationInstallations\Events\ApplicationInstallationContactPersonLinkedEvent;
-use Bitrix24\SDK\Application\Contracts\Bitrix24Accounts\Entity\Bitrix24AccountStatus;
-use Bitrix24\SDK\Application\Contracts\ContactPersons\Entity\ContactPersonInterface;
-use Bitrix24\SDK\Application\Contracts\ContactPersons\Entity\ContactPersonStatus;
-use Bitrix24\SDK\Application\Contracts\ContactPersons\Events\ContactPersonCreatedEvent;
-use Bitrix24\SDK\Application\Contracts\ContactPersons\Events\ContactPersonEmailChangedEvent;
-use Bitrix24\SDK\Application\PortalLicenseFamily;
-use Bitrix24\SDK\Core\Credentials\Scope;
-use Bitrix24\SDK\Core\Exceptions\InvalidArgumentException;
-use Bitrix24\Lib\Tests\EntityManagerFactory;
-use Bitrix24\SDK\Application\Contracts\ApplicationInstallations\Exceptions\ApplicationInstallationNotFoundException;
-use PHPUnit\Framework\Attributes\CoversClass;
-use PHPUnit\Framework\Attributes\Test;
-use PHPUnit\Framework\TestCase;
-use Psr\Log\NullLogger;
-use Symfony\Component\EventDispatcher\Debug\TraceableEventDispatcher;
-use Symfony\Component\EventDispatcher\EventDispatcher;
-use Symfony\Component\Stopwatch\Stopwatch;
-use Symfony\Component\Uid\Uuid;
-use libphonenumber\PhoneNumberUtil;
-use libphonenumber\PhoneNumber;
-use Bitrix24\Lib\Tests\Functional\ContactPersons\Builders\ContactPersonBuilder;
-use Bitrix24\Lib\ContactPersons\Enum\ContactPersonType;
-
-/**
- * @internal
- */
-#[CoversClass(Handler::class)]
-class HandlerTest extends TestCase
-{
-    private Handler $handler;
-
-    private Flusher $flusher;
-
-    private ContactPersonRepository $repository;
-
-    private ApplicationInstallationRepository $applicationInstallationRepository;
-
-    private Bitrix24AccountRepository $bitrix24accountRepository;
-
-    private TraceableEventDispatcher $eventDispatcher;
-
-    #[\Override]
-    protected function setUp(): void
-    {
-        $entityManager = EntityManagerFactory::get();
-        $this->eventDispatcher = new TraceableEventDispatcher(new EventDispatcher(), new Stopwatch());
-        $this->repository = new ContactPersonRepository($entityManager);
-        $this->applicationInstallationRepository = new ApplicationInstallationRepository($entityManager);
-        $this->bitrix24accountRepository = new Bitrix24AccountRepository($entityManager);
-        $this->flusher = new Flusher($entityManager, $this->eventDispatcher);
-        $this->handler = new Handler(
-            $this->applicationInstallationRepository,
-            $this->repository,
-            $this->flusher,
-            new NullLogger()
-        );
-    }
-
-    /**
-     * @throws InvalidArgumentException|\Random\RandomException
-     */
-    #[Test]
-    public function testInstallPartnerContactPersonSuccess(): void
-    {
-        // Подготовка Bitrix24 аккаунта и установки приложения
-        $applicationToken = Uuid::v7()->toRfc4122();
-        $memberId = Uuid::v4()->toRfc4122();
-        $externalId = Uuid::v7()->toRfc4122();
-
-        $bitrix24Account = (new Bitrix24AccountBuilder())
-            ->withApplicationScope(new Scope(['crm']))
-            ->withStatus(Bitrix24AccountStatus::new)
-            ->withApplicationToken($applicationToken)
-            ->withMemberId($memberId)
-            ->withMaster(true)
-            ->withSetToken()
-            ->withInstalled()
-            ->build();
-
-        $this->bitrix24accountRepository->save($bitrix24Account);
-
-        $applicationInstallation = (new ApplicationInstallationBuilder())
-            ->withApplicationStatus(new ApplicationStatus('F'))
-            ->withPortalLicenseFamily(PortalLicenseFamily::free)
-            ->withBitrix24AccountId($bitrix24Account->getId())
-            ->withApplicationStatusInstallation(ApplicationInstallationStatus::active)
-            ->withApplicationToken($applicationToken)
-            ->withContactPersonId(null)
-            ->withBitrix24PartnerContactPersonId(null)
-            ->withExternalId($externalId)
-            ->build();
-
-        $this->applicationInstallationRepository->save($applicationInstallation);
-        $this->flusher->flush();
-
-        // Данные контакта
-        $contactPersonBuilder = new ContactPersonBuilder();
-        $contactPerson = $contactPersonBuilder
-            ->withEmail('john.doe@example.com')
-            ->withMobilePhoneNumber($this->createPhoneNumber('+79991234567'))
-            ->withComment('Test comment')
-            ->withExternalId($externalId)
-            ->withBitrix24UserId($bitrix24Account->getBitrix24UserId())
-            ->withBitrix24PartnerId($applicationInstallation->getBitrix24PartnerId() ?? Uuid::v7())
-            ->build();
-
-        // Запуск use-case
-        $this->handler->handle(
-            new Command(
-                $applicationInstallation->getId(),
-                $contactPerson->getFullName(),
-                $bitrix24Account->getBitrix24UserId(),
-                $contactPerson->getUserAgentInfo(),
-                $contactPerson->getEmail(),
-                $contactPerson->getMobilePhone(),
-                $contactPerson->getComment(),
-                $contactPerson->getExternalId(),
-                $contactPerson->getBitrix24PartnerId(),
-            )
-        );
-
-        // Проверки: событие, связь и наличие контакта
-        $dispatchedEvents = $this->eventDispatcher->getOrphanedEvents();
-        $this->assertContains(ContactPersonCreatedEvent::class, $dispatchedEvents);
-        $this->assertContains(ApplicationInstallationBitrix24PartnerContactPersonLinkedEvent::class, $dispatchedEvents);
-
-        // Перечитаем установку и проверим привязку контактного лица (без поиска по externalId)
-        $foundInstallation = $this->applicationInstallationRepository->getById($applicationInstallation->getId());
-        $this->assertNotNull($foundInstallation->getBitrix24PartnerContactPersonId());
-
-        $uuid = $foundInstallation->getBitrix24PartnerContactPersonId();
-        $foundContactPerson = $this->repository->getById($uuid);
-        $this->assertInstanceOf(ContactPersonInterface::class, $foundContactPerson);
-        $this->assertEquals($foundContactPerson->getId(), $uuid);
-    }
-
-    #[Test]
-    public function testInstallPartnerContactPersonWithWrongApplicationInstallationId(): void
-    {
-        // Подготовим входные данные контакта (без реальной установки)
-        $contactPersonBuilder = new ContactPersonBuilder();
-        $contactPerson = $contactPersonBuilder
-            ->withEmail('john.doe@example.com')
-            ->withMobilePhoneNumber($this->createPhoneNumber('+79991234567'))
-            ->withComment('Test comment')
-            ->withExternalId(Uuid::v7()->toRfc4122())
-            ->build();
-
-        $uuidV7 = Uuid::v7();
-
-        $this->expectException(ApplicationInstallationNotFoundException::class);
-
-        $this->handler->handle(
-            new Command(
-                $uuidV7,
-                $contactPerson->getFullName(),
-                random_int(1, 1_000_000),
-                $contactPerson->getUserAgentInfo(),
-                $contactPerson->getEmail(),
-                $contactPerson->getMobilePhone(),
-                $contactPerson->getComment(),
-                $contactPerson->getExternalId(),
-                $contactPerson->getBitrix24PartnerId(),
-            )
-        );
-    }
-
-    private function createPhoneNumber(string $number): PhoneNumber
-    {
-        $phoneNumberUtil = PhoneNumberUtil::getInstance();
-        return $phoneNumberUtil->parse($number, 'RU');
-    }
-}
